[libc][rtlibc] remove libc_ioctl.h
