+++ include/rtlibc.h
@@ -16,7 +16,6 @@
 #include "libc/libc_errno.h"
 
 #include "libc/libc_fcntl.h"
-#include "libc/libc_ioctl.h"
 #include "libc/libc_dirent.h"
 #include "libc/libc_signal.h"
 #include "libc/libc_fdset.h"
